Add spacing on the left of main
Add space between the left edge of main and its content.

Add margin left to button and paragraph and add padding left
to inputs. The reason padding cannot be just added to main is because
it will push inputs to the left as well and this will not align with
the design.