Update height unit of body and html
Update the unit used for the height of the body and html from % to vh
because for some reason % doesn't extend to the whole page when the body
has little content. But, I believe this happens because usually when
percentages (%) are used we are referring to the container. So, if we
set:

main { height: 100% }

main will become 100% of its container's height. So, if the container's
height is 500px, then the main's height will be set to 100%.

But in this case, html doesn't have any container/parent so 100% is
pretty meaningless to it, I think.

Remove height from main as it is unnecessary.